Added long cast to avoid type mismatch errors.
Compiler error was: Type mismatch: cannot convert from Integer to long
See https://stackoverflow.com/questions/27105299 for more details

+++ src/net/ssehub/kernel_haven/undertaker/UndertakerExtractor.java
@@ -49,7 +49,7 @@ public class UndertakerExtractor extends AbstractCodeModelExtractor {
         fuzzyBooleanParsing = config.getValue(DefaultSettings.FUZZY_PARSING);
         
         config.registerSetting(HANG_TIMEOUT);
-        long timeout = config.getValue(HANG_TIMEOUT);
+        long timeout = (long) config.getValue(HANG_TIMEOUT);
         
         try {
             wrapper = new UndertakerWrapper(resourceDir, linuxSourceTree, timeout);
